requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
gunicorn
dj-database-url
psycopg2-binary
whitenoise
